Support multiple joysticks
Support multiple joysticks by saving the ID of the touch point when a
canvas is touched. Then when a touch is ended anywhere in the document,
each joystick compares the ended touch ID with the touch ID associated
with its canvas to know whether to honor that touch end event.